Make `PartStore` context-generic and move into its own module
