+++ code/actors/src/util/mod.rs
@@ -1,7 +1,9 @@
 mod forward;
 mod make_actor;
+mod part_store;
 pub mod value_builder;
 
 pub use forward::{forward, Forward};
 pub use make_actor::make_node_actor;
+pub use part_store::PartStore;
 pub use value_builder::{test::TestValueBuilder, ValueBuilder};
